refactor(dashboard): migrate Usage panel + uPlot loader
Stage 4 PR 4.3 from #28. Usage panel pulls out with its uPlot
chart loader and the chart subcomponent.

- panels/usage.ts (231 LoC) — UsagePanel (renamed from
  UsageWithChart), UsageChart subcomponent, loadUPlot helper
  (private). Daily series chart, rolling-window table, top
  models — same as before.
- dashboard/app.js — drops the live UsageWithChart + UsageChart
  + loadUPlot (~210 LoC) AND the dead-code original UsagePanel
  (~85 LoC) that hadn't been wired since the chart rewrite.
  TABS entry points at the new UsagePanel symbol.

Net dashboard/app.js loses 295 LoC; bundle still ~122 KB. uPlot
import stays external (esm.sh URL preserved through tsup).
1682 tests still pass.
