+++ dashboard/app.js
@@ -34,97 +34,11 @@ import { PlansPanel } from "./src/panels/plans";
 import { SettingsPanel } from "./src/panels/settings";
 import { SkillsPanel } from "./src/panels/skills";
 import { ToolsPanel } from "./src/panels/tools";
+import { UsagePanel } from "./src/panels/usage";
 
 const html = htm.bind(h);
 
 // ---------- panels ----------
-
-function UsagePanel() {
-  const { data, error, loading } = usePoll("/usage", 5000);
-  if (loading && !data) return html`<div class="boot">loading usage…</div>`;
-  if (error) return html`<div class="notice err">usage failed: ${error.message}</div>`;
-  const u = data;
-
-  return html`
-    <div>
-      <div class="panel-header">
-        <h2 class="panel-title">Usage</h2>
-        <span class="panel-subtitle">${u.recordCount.toLocaleString()} records · ${u.logSize}</span>
-      </div>
-
-      ${
-        u.recordCount === 0
-          ? html`<div class="empty">No usage data yet — run a turn in <code>reasonix chat</code> / <code>code</code> / <code>run</code> and refresh.</div>`
-          : html`
-          <table>
-            <thead>
-              <tr>
-                <th></th>
-                <th class="numeric">turns</th>
-                <th class="numeric">cache hit</th>
-                <th class="numeric">cost (USD)</th>
-                <th class="numeric">cache saved</th>
-                <th class="numeric">vs Claude</th>
-                <th class="numeric">saved</th>
-              </tr>
-            </thead>
-            <tbody>
-              ${u.buckets.map((b) => {
-                const hitRatio =
-                  b.cacheHitTokens + b.cacheMissTokens > 0
-                    ? b.cacheHitTokens / (b.cacheHitTokens + b.cacheMissTokens)
-                    : 0;
-                const claudeSavings = b.claudeEquivUsd > 0 ? 1 - b.costUsd / b.claudeEquivUsd : 0;
-                return html`
-                  <tr>
-                    <td>${b.label}</td>
-                    <td class="numeric">${fmtNum(b.turns)}</td>
-                    <td class="numeric">${b.turns > 0 ? fmtPct(hitRatio) : "—"}</td>
-                    <td class="numeric">${b.turns > 0 ? fmtUsd(b.costUsd) : "—"}</td>
-                    <td class="numeric">${b.turns > 0 && b.cacheSavingsUsd > 0 ? fmtUsd(b.cacheSavingsUsd) : "—"}</td>
-                    <td class="numeric">${b.turns > 0 ? fmtUsd(b.claudeEquivUsd) : "—"}</td>
-                    <td class="numeric">${b.turns > 0 && claudeSavings > 0 ? fmtPct(claudeSavings) : "—"}</td>
-                  </tr>
-                `;
-              })}
-            </tbody>
-          </table>
-        `
-      }
-
-      ${
-        u.byModel.length > 0
-          ? html`
-          <div class="section-title">Most used models</div>
-          <table>
-            <thead><tr><th>model</th><th class="numeric">turns</th></tr></thead>
-            <tbody>
-              ${u.byModel.slice(0, 5).map(
-                (m) => html`
-                <tr><td><code>${m.model}</code></td><td class="numeric">${fmtNum(m.turns)}</td></tr>
-              `,
-              )}
-            </tbody>
-          </table>
-        `
-          : null
-      }
-
-      ${
-        u.subagents
-          ? html`
-          <div class="section-title">Subagent activity</div>
-          <div class="card">
-            <div class="card-title">Total runs</div>
-            <div class="card-value">${fmtNum(u.subagents.total)}</div>
-            <div class="card-hint">${fmtUsd(u.subagents.costUsd)} · ${(u.subagents.totalDurationMs / 1000).toFixed(1)}s total</div>
-          </div>
-        `
-          : null
-      }
-    </div>
-  `;
-}
 
 function renderMessageBody(text) {
   if (!text) return null;
@@ -1568,230 +1482,6 @@ function SessionsPanel() {
 
 // ---------- Plans archive ----------
 
-// ---------- Usage time-series chart (uPlot) ----------
-
-let uPlotPromise = null;
-function loadUPlot() {
-  if (!uPlotPromise) {
-    uPlotPromise = import("https://esm.sh/uplot@1.6.31").then((m) => m.default ?? m);
-  }
-  return uPlotPromise;
-}
-
-function UsageChart({ days }) {
-  const containerRef = useRef(null);
-  const plotRef = useRef(null);
-
-  useEffect(() => {
-    let cancelled = false;
-    loadUPlot().then((uPlot) => {
-      if (cancelled || !containerRef.current) return;
-      // Destroy previous instance on data refresh.
-      if (plotRef.current) {
-        plotRef.current.destroy();
-        plotRef.current = null;
-      }
-      // Don't render an empty chart — let the parent show a fallback.
-      if (!days || days.length === 0) return;
-      const xs = days.map((d) => Math.floor(Date.parse(d.day) / 1000));
-      const cost = days.map((d) => d.costUsd);
-      const saved = days.map((d) => d.cacheSavingsUsd);
-      const turns = days.map((d) => d.turns);
-      const data = [xs, cost, saved, turns];
-      const opts = {
-        width: containerRef.current.clientWidth,
-        height: 280,
-        cursor: { drag: { x: true, y: false } },
-        scales: {
-          x: { time: true },
-          y: { auto: true },
-          turns: { auto: true },
-        },
-        axes: [
-          {
-            stroke: "#94a3b8",
-            grid: { stroke: "rgba(148, 163, 184, 0.08)" },
-          },
-          {
-            scale: "y",
-            label: "USD",
-            stroke: "#94a3b8",
-            grid: { stroke: "rgba(148, 163, 184, 0.08)" },
-            values: (_u, v) => v.map((n) => `$${n.toFixed(4)}`),
-          },
-          {
-            scale: "turns",
-            side: 1,
-            label: "turns",
-            stroke: "#94a3b8",
-            grid: { show: false },
-          },
-        ],
-        series: [
-          {},
-          {
-            label: "cost",
-            stroke: "#67e8f9",
-            width: 2,
-            fill: "rgba(103, 232, 249, 0.10)",
-          },
-          {
-            label: "cache saved",
-            stroke: "#5eead4",
-            width: 2,
-            dash: [4, 4],
-          },
-          {
-            label: "turns",
-            stroke: "#c4b5fd",
-            scale: "turns",
-            width: 1.5,
-            points: { show: true, size: 4 },
-          },
-        ],
-        legend: { live: true },
-      };
-      plotRef.current = new uPlot(opts, data, containerRef.current);
-    });
-
-    // Resize observer keeps the chart at full panel width.
-    const ro = new ResizeObserver(() => {
-      if (plotRef.current && containerRef.current) {
-        plotRef.current.setSize({
-          width: containerRef.current.clientWidth,
-          height: 280,
-        });
-      }
-    });
-    if (containerRef.current) ro.observe(containerRef.current);
-
-    return () => {
-      cancelled = true;
-      ro.disconnect();
-      if (plotRef.current) {
-        plotRef.current.destroy();
-        plotRef.current = null;
-      }
-    };
-  }, [days]);
-
-  return html`<div ref=${containerRef} style="width: 100%; min-height: 280px;"></div>`;
-}
-
-// ---------- existing UsagePanel rewrite — chart + table ----------
-
-function UsageWithChart() {
-  const { data: summary, error, loading } = usePoll("/usage", 5000);
-  const [series, setSeries] = useState(null);
-
-  useEffect(() => {
-    let cancelled = false;
-    (async () => {
-      try {
-        const s = await api("/usage/series");
-        if (!cancelled) setSeries(s.days ?? []);
-      } catch {
-        /* keep null; chart hides */
-      }
-    })();
-    const t = setInterval(async () => {
-      try {
-        const s = await api("/usage/series");
-        if (!cancelled) setSeries(s.days ?? []);
-      } catch {
-        /* swallow */
-      }
-    }, 30_000);
-    return () => {
-      cancelled = true;
-      clearInterval(t);
-    };
-  }, []);
-
-  if (loading && !summary) return html`<div class="boot">loading usage…</div>`;
-  if (error) return html`<div class="notice err">usage failed: ${error.message}</div>`;
-  const u = summary;
-
-  return html`
-    <div>
-      <div class="panel-header">
-        <h2 class="panel-title">Usage</h2>
-        <span class="panel-subtitle">${u.recordCount.toLocaleString()} records · ${u.logSize}</span>
-      </div>
-
-      ${
-        series && series.length > 0
-          ? html`
-          <div class="card" style="padding: 18px;">
-            <div class="card-title" style="margin-bottom: 12px;">Daily usage (cost · cache saved · turns)</div>
-            <${UsageChart} days=${series} />
-          </div>
-        `
-          : null
-      }
-
-      ${
-        u.recordCount === 0
-          ? html`<div class="empty" style="margin-top: 16px;">No usage data yet — run a turn in <code>reasonix chat</code> / <code>code</code> / <code>run</code> and refresh.</div>`
-          : html`
-          <div class="section-title">Rolling windows</div>
-          <table>
-            <thead>
-              <tr>
-                <th></th>
-                <th class="numeric">turns</th>
-                <th class="numeric">cache hit</th>
-                <th class="numeric">cost (USD)</th>
-                <th class="numeric">cache saved</th>
-                <th class="numeric">vs Claude</th>
-                <th class="numeric">saved</th>
-              </tr>
-            </thead>
-            <tbody>
-              ${u.buckets.map((b) => {
-                const hitRatio =
-                  b.cacheHitTokens + b.cacheMissTokens > 0
-                    ? b.cacheHitTokens / (b.cacheHitTokens + b.cacheMissTokens)
-                    : 0;
-                const claudeSavings = b.claudeEquivUsd > 0 ? 1 - b.costUsd / b.claudeEquivUsd : 0;
-                return html`
-                  <tr>
-                    <td>${b.label}</td>
-                    <td class="numeric">${fmtNum(b.turns)}</td>
-                    <td class="numeric">${b.turns > 0 ? fmtPct(hitRatio) : "—"}</td>
-                    <td class="numeric">${b.turns > 0 ? fmtUsd(b.costUsd) : "—"}</td>
-                    <td class="numeric">${b.turns > 0 && b.cacheSavingsUsd > 0 ? fmtUsd(b.cacheSavingsUsd) : "—"}</td>
-                    <td class="numeric">${b.turns > 0 ? fmtUsd(b.claudeEquivUsd) : "—"}</td>
-                    <td class="numeric">${b.turns > 0 && claudeSavings > 0 ? fmtPct(claudeSavings) : "—"}</td>
-                  </tr>
-                `;
-              })}
-            </tbody>
-          </table>
-        `
-      }
-
-      ${
-        u.byModel.length > 0
-          ? html`
-          <div class="section-title">Most used models</div>
-          <table>
-            <thead><tr><th>model</th><th class="numeric">turns</th></tr></thead>
-            <tbody>
-              ${u.byModel.slice(0, 5).map(
-                (m) => html`
-                <tr><td><code>${m.model}</code></td><td class="numeric">${fmtNum(m.turns)}</td></tr>
-              `,
-              )}
-            </tbody>
-          </table>
-        `
-          : null
-      }
-    </div>
-  `;
-}
-
 // ---------- Semantic index ----------
 
 function SemanticPanel() {
@@ -2356,7 +2046,7 @@ const TABS = [
     id: "usage",
     name: "Usage",
     glyph: "$",
-    panel: () => html`<${UsageWithChart} />`,
+    panel: () => html`<${UsagePanel} />`,
     ready: true,
     badge: null,
   },
